Reorganize files in a more logical fashion
